% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjustRaster.R
\name{adjustRaster}
\alias{adjustRaster}
\title{Adjust Raster Data}
\usage{
adjustRaster(raster_data, correction_type)
}
\arguments{
\item{raster_data}{Raster data to correct. This can be a file path to a single raster file,
a file path to multiple raster files (to be combined into a stack), a RasterLayer object,
or a RasterStack object.}

\item{correction_type}{The type of correction to apply ("reflectance" or "temperature").}
}
\value{
A RasterLayer or RasterStack object with the correction applied.
}
\description{
This function takes raster data, applies a scale factor and offset correction,
and returns the corrected raster layer. It is possible to choose between "reflectance"
or "temperature" correction types.
}
\examples{
# corrected_raster_reflectance <- adjustRaster("path_to_raster.tif", "reflectance")
# plot(corrected_raster_reflectance)
# corrected_raster_temperature <- adjustRaster("path_to_raster.tif", "temperature")
# plot(corrected_raster_temperature)

}
