% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indexplot.R
\name{indexplot}
\alias{indexplot}
\title{Index Plot: Smooth Scatter Plot for LST and Indices}
\usage{
indexplot(
  lst,
  index,
  index_name,
  main = NULL,
  col_alpha = 0.15,
  cex_lab = 1.2,
  cex_main = 1.2,
  cex_axis = 1.2
)
}
\arguments{
\item{lst}{Raster Layer of LST values.}

\item{index}{Raster Layer of index values (NDVI, NDBI, etc.).}

\item{index_name}{Character string with the name of the index ("NDVI", "NDBI", etc.) for labeling.}

\item{main}{Title for the plot, automatically generated if NULL.}

\item{col_alpha}{Transparency level for the plot points.}

\item{cex_lab}{Magnification for x and y labels relative to the current setting of cex.}

\item{cex_main}{Magnification for the main title relative to the current setting of cex.}

\item{cex_axis}{Magnification for axis text relative to the current setting of cex.}
}
\value{
A Plot
}
\description{
Creates a scatter plot with a smoothed line for LST values against another index like NDVI or NDBI.
}
\examples{

  #indexplot(lst_raster, ndvi_raster, "NDVI")

}
