% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NDBI.R
\name{NDBI}
\alias{NDBI}
\title{Computes the NDBI using NIR and the SWIR band.}
\usage{
NDBI(swir, nir)
}
\arguments{
\item{swir}{Raster Layer of the Short-Wave Infrared band.}

\item{nir}{Raster Layer of the Near-Infrared band.}
}
\value{
Raster Layer containing NDBI values.
}
\description{
Computes the NDBI using NIR and the SWIR band.
}
\examples{
# ndbi_calc <- NDBI (swir_band, nir_band)
# plot(ndbi_calc)
}
