% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NDVI.R
\name{NDVI}
\alias{NDVI}
\title{Calculate Normalized Difference Vegetation Index (NDVI)}
\usage{
NDVI(nir, red)
}
\arguments{
\item{nir}{Raster Layer of the Near-Infrared band.}

\item{red}{Raster Layer of the Red band.}
}
\value{
Raster Layer containing NDVI values.
}
\description{
Computes the NDVI from NIR and Red bands.
}
\examples{
#ndvi_calc <- NDVI(nir_band, red_band)
#plot(ndvi_calc)

}
