% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rural_area_extraction.R
\name{rural_area_extraction}
\alias{rural_area_extraction}
\title{Calculate External Area of City Buffer}
\usage{
rural_area_extraction(city_geom, dist)
}
\arguments{
\item{city_geom}{A spatial object representing the city boundaries.}

\item{dist}{The distance for the buffer around the city, in the same units as the city geometry.}
}
\value{
A spatial object representing the external area of the buffer.
}
\description{
This function creates a buffer around a city geometry and then calculates
the external area by subtracting the original geometry from the buffer.
}
\examples{

#external_area <- calculate_external_buffer(city_geom, 2000)
#plot(st_geometry(external_area), col = "black", main = "External area of the buffer")
}
