% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trimRaster.R
\name{trimRaster}
\alias{trimRaster}
\title{Trim Raster}
\usage{
trimRaster(r_stack, polygon)
}
\arguments{
\item{r_stack}{RasterLayer or RasterStack object to be cropped and masked.}

\item{polygon}{RasterLayer object used for cropping and masking.}
}
\value{
Raster object that has been cropped and masked.
}
\description{
This function applies both crop and mask operations on a raster object.
}
\examples{
# trimmed_raster <- trimRaster(raster_object, mask_object)
}
